Move PdfOpener to ui-core module
COAND-366
