+++ online-banking-core/src/main/java/com/adyen/checkout/onlinebankingcore/internal/ui/DefaultOnlineBankingDelegate.kt
@@ -29,7 +29,6 @@ import com.adyen.checkout.core.internal.util.Logger
 import com.adyen.checkout.onlinebankingcore.internal.ui.model.OnlineBankingInputData
 import com.adyen.checkout.onlinebankingcore.internal.ui.model.OnlineBankingModel
 import com.adyen.checkout.onlinebankingcore.internal.ui.model.OnlineBankingOutputData
-import com.adyen.checkout.onlinebankingcore.internal.util.PdfOpener
 import com.adyen.checkout.onlinebankingcore.internal.util.getLegacyIssuers
 import com.adyen.checkout.onlinebankingcore.internal.util.mapToModel
 import com.adyen.checkout.ui.core.internal.ui.ButtonComponentViewType
@@ -37,6 +36,7 @@ import com.adyen.checkout.ui.core.internal.ui.ComponentViewType
 import com.adyen.checkout.ui.core.internal.ui.PaymentComponentUIEvent
 import com.adyen.checkout.ui.core.internal.ui.PaymentComponentUIState
 import com.adyen.checkout.ui.core.internal.ui.SubmitHandler
+import com.adyen.checkout.ui.core.internal.util.PdfOpener
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.Flow
