+++ ui-core/src/main/java/com/adyen/checkout/ui/core/internal/util/PdfOpener.kt
@@ -6,7 +6,7 @@
  * Created by atef on 20/9/2022.
  */
 
-package com.adyen.checkout.onlinebankingcore.internal.util
+package com.adyen.checkout.ui.core.internal.util
 
 import android.content.ActivityNotFoundException
 import android.content.Context
@@ -18,7 +18,6 @@ import androidx.browser.customtabs.CustomTabColorSchemeParams
 import androidx.browser.customtabs.CustomTabsIntent
 import com.adyen.checkout.core.internal.util.LogUtil
 import com.adyen.checkout.core.internal.util.Logger
-import com.adyen.checkout.ui.core.internal.util.ThemeUtil
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 class PdfOpener {
